[layout] Remove dry-run layout flag.
Dry-run flag was intended for figuring out whether an element is going to be block or inline,
needed during block container layout(because we need to know if the block container only has block children,
inline children, or both).
But this "feature" made writing debug logs much harder, due to a lot of meaningless log outputs.

Now instead we have isElementBlockLevel, a function dedicated for this job.