[yw_common] yw_duplicate_str: don't use yw_append_str to create string
yw_append_str had special case where if dest pointer points to NULL, it returns duplication of string,
instead of appending to something. This was (and is) useful for building strings.

And for a long time, yw_duplicate_str depended on this behavior, but that special logic is exactly
what yw_duplicate_str should be doing, so it makes sense to move that logic to yw_duplicate_str.
(and make yw_append_str call yw_duplicate_str instead)